refactored tests; vendored in sqlite3; added sqlite persistence (via string store)

+++ src/padamose/gtest/TestVariant.cpp
@@ -25,57 +25,57 @@ TEST ( Variant, test_coercion ) {
     v = true;
     ASSERT_TRUE ( v.get < bool >() == true );
     ASSERT_TRUE ( v.get < double >() == 1.0 );
-    ASSERT_TRUE ( v.get < int >() == 1 );
-    ASSERT_TRUE ( v.get < size_t >() == 1 );
+    ASSERT_TRUE ( v.get < s64 >() == 1 );
+    ASSERT_TRUE ( v.get < u64 >() == 1 );
     ASSERT_TRUE ( v.get < string >() == "true" );
     
     v = false;
     ASSERT_TRUE ( v.get < bool >() == false );
     ASSERT_TRUE ( v.get < double >() == 0.0 );
-    ASSERT_TRUE ( v.get < int >() == 0 );
-    ASSERT_TRUE ( v.get < size_t >() == 0 );
+    ASSERT_TRUE ( v.get < s64 >() == 0 );
+    ASSERT_TRUE ( v.get < u64 >() == 0 );
     ASSERT_TRUE ( v.get < string >() == "false" );
     
     v = 123.0;
     ASSERT_TRUE ( v.get < bool >() == true );
     ASSERT_TRUE ( v.get < double >() == 123.0 );
-    ASSERT_TRUE ( v.get < int >() == 123 );
-    ASSERT_TRUE ( v.get < size_t >() == 123 );
+    ASSERT_TRUE ( v.get < s64 >() == 123 );
+    ASSERT_TRUE ( v.get < u64 >() == 123 );
     ASSERT_TRUE ( v.get < string >() == "123" );
     
     v = 0.0;
     ASSERT_TRUE ( v.get < bool >() == false );
     ASSERT_TRUE ( v.get < double >() == 0.0 );
-    ASSERT_TRUE ( v.get < int >() == 0 );
-    ASSERT_TRUE ( v.get < size_t >() == 0 );
+    ASSERT_TRUE ( v.get < s64 >() == 0 );
+    ASSERT_TRUE ( v.get < u64 >() == 0 );
     ASSERT_TRUE ( v.get < string >() == "0" );
     
-    v = 456;
+    v = ( s64 )456;
     ASSERT_TRUE ( v.get < bool >() == true );
     ASSERT_TRUE ( v.get < double >() == 456.0 );
-    ASSERT_TRUE ( v.get < int >() == 456 );
-    ASSERT_TRUE ( v.get < size_t >() == 456 );
+    ASSERT_TRUE ( v.get < s64 >() == 456 );
+    ASSERT_TRUE ( v.get < u64 >() == 456 );
     ASSERT_TRUE ( v.get < string >() == "456" );
     
-    v = 0;
+    v = ( s64 )0;
     ASSERT_TRUE ( v.get < bool >() == false );
     ASSERT_TRUE ( v.get < double >() == 0.0 );
-    ASSERT_TRUE ( v.get < int >() == 0 );
-    ASSERT_TRUE ( v.get < size_t >() == 0 );
+    ASSERT_TRUE ( v.get < s64 >() == 0 );
+    ASSERT_TRUE ( v.get < u64 >() == 0 );
     ASSERT_TRUE ( v.get < string >() == "0" );
     
     v = "123";
     ASSERT_TRUE ( v.get < bool >() == true );
     ASSERT_TRUE ( v.get < double >() == 123.0 );
-    ASSERT_TRUE ( v.get < int >() == 123 );
-    ASSERT_TRUE ( v.get < size_t >() == 123 );
+    ASSERT_TRUE ( v.get < s64 >() == 123 );
+    ASSERT_TRUE ( v.get < u64 >() == 123 );
     ASSERT_TRUE ( v.get < string >() == "123" );
     
     v = string ( "123" );
     ASSERT_TRUE ( v.get < bool >() == true );
     ASSERT_TRUE ( v.get < double >() == 123.0 );
-    ASSERT_TRUE ( v.get < int >() == 123 );
-    ASSERT_TRUE ( v.get < size_t >() == 123 );
+    ASSERT_TRUE ( v.get < s64 >() == 123 );
+    ASSERT_TRUE ( v.get < u64 >() == 123 );
     ASSERT_TRUE ( v.get < string >() == "123" );
 }
